requirements.txt
langchain
python-dotenv
streamlit
PyPDF2
sentence-transformers